pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomous-dev-env"
version = "0.1.0"
description = "AI-Native Cycle-Based Contract-Driven Development Environment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic~=2.11.0",
    "python-dotenv",
    "pydantic-settings",
    "pydantic-ai>=0.0.18",
    "langgraph>=1.0.5",
    "e2b-code-interpreter>=2.4.1",
    "httpx>=0.28.1",
    "google-auth>=2.43.0",
    "litellm>=1.80.11",
    "typer",
    "rich",
    "pyyaml",
    "ase",
]

[tool.uv]
environments = ["sys_platform == 'linux'"]

[project.scripts]
ac-cdd = "ac_cdd_core.cli:app"
mlip-autopipec = "mlip_autopipec.cli.main:app"

[dependency-groups]
dev = [
    "shellingham",
    "pytest",
    "pytest-cov",
    "hypothesis",
    "playwright",
    "vcrpy",
    "ruff",
    "google-genai",
    "mypy",
    "bandit",
    "types-PyYAML",
    "pytest-asyncio>=1.3.0",
    "pre-commit>=4.5.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
# E/F/I: Basic, B: Bugbear, S: Bandit, UP: pyupgrade
select = [
    "F",    # Pyflakes (basic errors)
    "E", "W", # pycodestyle (style)
    "C90",  # mccabe (complexity check: prevent AI spaghetti code)
    "I",    # isort (import organization)
    "N",    # pep8-naming (naming conventions: proper class/function names)
    "UP",   # pyupgrade (modern syntax)
    "YTT",  # flake8-2020 (prevent sys.version misuse)
    "ANN",  # flake8-annotations (enforce type hints: critical)
    "ASYNC",# flake8-async (prevent async bugs)
    "S",    # flake8-bandit (security)
    "BLE",  # flake8-blind-except (prohibit catching all errors)
    "B",    # flake8-bugbear (detect bug hotspots)
    "A",    # flake8-builtins (prevent overwriting built-in names: id, list, etc.)
    "C4",   # flake8-comprehensions (optimize list comprehensions)
    "DTZ",  # flake8-datetimez (enforce timezone)
    "T10",  # flake8-debugger (prevent forgotten debugger statements)
    "EM",   # flake8-errmsg (improve exception message quality)
    "ICN",  # flake8-import-conventions (standardize import aliases: pd, np, etc.)
    "PIE",  # flake8-pie (miscellaneous code improvements)
    "T20",  # flake8-print (prohibit print statements: enforce logger usage)
    "PT",   # flake8-pytest-style (improve test code quality)
    "RET",  # flake8-return (return statement consistency)
    "SIM",  # flake8-simplify (concise writing)
    "TID",  # flake8-tidy-imports (organize imports)
    "ARG",  # flake8-unused-arguments (remove unused arguments)
    "PTH",  # flake8-use-pathlib (prohibit os.path -> enforce pathlib)
    "ERA",  # eradicate (remove commented code: clean up AI trial-and-error traces)
    "PL",   # Pylint (comprehensive code quality checks)
    "TRY",  # tryceratops (proper exception handling)
    "RUF",  # Ruff-specific rules
]
# S101: Allow use of assert (for tests)
ignore = [
    "ANN101", # Type hint for self is unnecessary (deprecated)
    "ANN002", # Type hint for *args can be omitted
    "ANN003", # Type hint for **kwargs can be omitted
    "ANN001", # Missing type annotation for function arguments
    "ANN401", # Dynamically typed expressions are disallowed
    "ARG001", # Unused function argument
    "ARG005", # Unused lambda argument
    "PLR2004", # Magic value used in comparison
    "E501",   # Line length limit (auto-fix may not work, slows development)
    "TRY003", # Exception message too long warning (too strict for AI)
    "D",      # docstring (excluded to focus on code behavior risks. Add if needed)
    "ANN201", # Missing return type annotation for public functions
    "N806",   # Variable should be lowercase
    "PLC0415", # Import should be at top-level
    "BLE001", # Blind exception catch
    "PT019",  # Fixture without value injection
    "RUF003", # Ambiguous multiplication sign
    "ARG002", # Unused method argument
    "RUF043", # Pattern metacharacters not escaped in match
    "RUF059", # Unused unpacked variable
    "PLR0913", # Too many arguments in function definition
    "ANN202"  # Missing return type annotation for private functions
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.lint.mccabe]
max-complexity = 10 # Exceeding this prompts function splitting (prevent AI long functions)

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true # Arguments starting with _ do not require type hints


[tool.pytest.ini_options]
addopts = "--cov=dev_src --cov=src --cov-report=term-missing"
testpaths = ["tests"]


[tool.mypy]
strict = true
ignore_missing_imports = true


[tool.hatch.build.targets.wheel]
packages = ["dev_src/ac_cdd_core", "src/mlip_autopipec"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
targets = ["dev_src"]
